rearranged files + moved navbar to bottom + made footer responsive
