+++ src/components/Body/body.css
@@ -1,0 +1,72 @@
+.main-body {
+  width: 65%;
+  background-color: #4b9fe0;
+  border: 4px solid #1a374d;
+  border-left: none;
+  border-bottom: none;
+}
+
+.site-img {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+}
+
+.main-body img {
+  width: 85%;
+  height: 30vw;
+  opacity: 0.8;
+  transition: 0.3s;
+  border: 5px solid #1a374d;
+}
+
+.main-body img:hover {
+  opacity: 1;
+}
+
+.site-space {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.site-links {
+  width: 90%;
+}
+
+.site-links button {
+  background-color: #406882;
+  color: aliceblue;
+  border-radius: 12px;
+}
+
+.site-description {
+  width: 95%;
+  margin-top: 2%;
+  background-color: #5a93b9;
+  border: 2px solid #1a374d;
+  padding: 1%;
+  text-align: center;
+}
+
+.site-description p {
+  font-size: 2rem;
+}
+
+.site-space h2 {
+  font-size: 3rem;
+  font-style: italic;
+  color: #224864;
+  text-shadow: black;
+}
+
+.bottom {
+  background-color: #4b9fe0;
+  height: 30vh;
+}
+
+@media only screen and (max-width: 850px) {
+  .main-body {
+    width: 100%;
+  }
+}
