+++ src/components/Footer/footer.css
@@ -1,0 +1,37 @@
+.contact-links {
+  display: flex;
+  justify-content: space-evenly;
+  width: 100%;
+  list-style-type: none;
+  /* For chromium based user agent stylesheet default left padding0 */
+  padding-left: 0;
+}
+
+.contact-links a {
+  text-decoration: none;
+  color: #406882;
+  font-weight: bold;
+  font-size: 27px;
+}
+
+.contact-links a:hover {
+  color: #1a374d;
+}
+
+@media only screen and (max-width: 1200px) {
+  /*Tablets [601px -> 1200px]*/
+}
+@media only screen and (max-width: 850px) {
+  .contact-links {
+    flex-direction: column;
+    align-items: center;
+    margin: 10px 0;
+  }
+  .contact-links li {
+    margin-bottom: 20px;
+  }
+  /*Big smartphones [426px -> 600px]*/
+}
+@media only screen and (max-width: 425px) {
+  /*Small smartphones [325px -> 425px]*/
+}
