+++ src/components/Footer/index.jsx
@@ -1,0 +1,28 @@
+import "./footer.css";
+
+function Footer() {
+  return (
+    <footer>
+      <nav className="footer">
+        <ul className="contact-links">
+          <li>
+            <a href="tel:+18455216432">(845) 521-6432 </a>
+          </li>
+          <li>
+            <a href="mailto:machlisj@gmail.com">machlisj@gmail.com</a>
+          </li>
+          <li>
+            <a href="https://github.com/Flumanuck">Github</a>
+          </li>
+          <li>
+            <a href="https://www.linkedin.com/in/jacob-machlis-729695159/">
+              LinkedIn
+            </a>
+          </li>
+        </ul>
+      </nav>
+    </footer>
+  );
+}
+
+export default Footer;
